fix(numpy): aliases of builtin types is deprecated

+++ flopy/plot/crosssection.py
@@ -631,7 +631,7 @@ class PlotCrossSection:
             else:
                 ibound = self.mg.idomain
 
-        plotarray = np.zeros(ibound.shape, dtype=np.int)
+        plotarray = np.zeros(ibound.shape, dtype=int)
         idx1 = ibound == 0
         plotarray[idx1] = 1
         plotarray = np.ma.masked_equal(plotarray, 0)
@@ -685,7 +685,7 @@ class PlotCrossSection:
 
             ibound = self.mg.idomain
 
-        plotarray = np.zeros(ibound.shape, dtype=np.int)
+        plotarray = np.zeros(ibound.shape, dtype=int)
         idx1 = ibound == 0
         idx2 = ibound < 0
         plotarray[idx1] = 1
@@ -823,11 +823,11 @@ class PlotCrossSection:
                     idx = mflist["node"]
 
         if len(self.mg.shape) != 3:
-            plotarray = np.zeros((self._nlay, self._ncpl), dtype=np.int)
+            plotarray = np.zeros((self._nlay, self._ncpl), dtype=int)
             plotarray[tuple(idx)] = 1
         else:
             plotarray = np.zeros(
-                (self.mg.nlay, self.mg.nrow, self.mg.ncol), dtype=np.int
+                (self.mg.nlay, self.mg.nrow, self.mg.ncol), dtype=int
             )
             plotarray[idx[0], idx[1], idx[2]] = 1
 
@@ -1155,7 +1155,7 @@ class PlotCrossSection:
             # thickness by setting laytyp to zeros
             if head is None or laytyp is None:
                 head = np.zeros(botm.shape, np.float32)
-                laytyp = np.zeros((nlay,), dtype=np.int)
+                laytyp = np.zeros((nlay,), dtype=int)
                 head[0, :, :] = top
                 if nlay > 1:
                     head[1:, :, :] = botm[:-1, :, :]
@@ -1170,7 +1170,7 @@ class PlotCrossSection:
             )
 
             if qz is None:
-                qz = np.zeros(qx.shape, dtype=np.float)
+                qz = np.zeros(qx.shape, dtype=float)
 
             qx = qx.ravel()
             qy = qy.ravel()
